Collapse ssa builder into env builder
Summary:
Let's stop pretending that we're extending the ssa builder. The only shared behavior we benefit from is the abrupt completion handling, but almost every ssa-like function is being overridden in our subclass.

Moreover, the fundamental data structures we operate over (Val.t and Ssa_api.t) in the SSA builder are not enough to describe the functionality required in the EnvBuilder.

Instead of functorizing and complicating this code more with ocaml type system features in the name of "code reuse", we should instead accept that we are not actually "extending" this class and just fold it into the EnvBuilder itself.

This diff gets rid of the extension and instead copy/pastes all the non-overridden methods directly into the EnvBuilder.

[Facebook]
Another benefit to this is that we no longer need to worry about breaking the ssa builder and the downstream effects of such a breakage.

Reviewed By: mvitousek

Differential Revision: D28811792

fbshipit-source-id: 390a3879abcbedcf37d5790ad06c171feb88a917