Model refinements as writes
Summary:
Now we model refinements as SSA writes. There were two primary challenges here:
1: The location of a refinement is not unique enough to identify a refinement. For example:
```
if (x) {
  x;
} else {
  x;
}
```

the `x` in the consequent and else branch are both refined by `if(x)`, but their refinements are not the same.

2: The scopes of refinements do not match the scope of runtime writes. For example:

```
(x != null);
x; // If we naively model x != null as a write then we would report that x on this line is not null, even though that is false
```

To fix the first issue, we swap out . The `Val`  module provided by the  ssa_builder with a new module specifically made for the EnvBuilder. The new `Refinement` `Val.write_state`  we introduce there _wraps_ a full Val.t. This allows us to express refinements over a write instead of having to make a location to represent the location of the refinement. This Refinement Val gets turned into an Env_api.Refinement. The Env_api is the Ssa_api equivalent for the Env builder, which we also introduce in this diff. We also expose a `refinement_of_id` function that can take the refinement_ids referenced in these types and turn them into `refinement_kind`s.

The `Refinement` `Val.write_state` also solves our second issue-- it is trivial to "restore" the unrefined value if we track refinements in this way. Notice the new implementations for pushing and popping refinement scopes -- we examine the scope and add/remove any refinements as necessary. This allows us to explicitly mark the scope of refinements.

Since we still model the latest_refinements as a stack, it is still simple to negate new refinements.

However, we have changed the type of the entries on the latest_refinement stack. Instead of mapping ssa ids -> refinement ids, we map variable names to an (ssa_id, refinement_id) pair. We use the refinement_id and ssa_id to ensure that we apply the correct refinement to the correct Val.ts when pushing/popping refinement scopes.

Other than the implementation details of the latest_refinement stack and what pushing/popping refinements actually does, the code that uses the refinement stack and pushes/pops is mostly unchanged. The primary difference is that we need to be more careful about merging the non-refined versions of environments (when necessary), and to support that I introduced the `ssa_env_without_refinements` function.

Zooming out a bit, the operations described here (pushing scopes, popping scopes) etc suffer from the same problems that the original ssa_builder's helper functions do: they are too fine-grained and require too much understanding to sequence these side-effectful functions properly. In a future diff, I plan to introduce higher level "control-flow" helper functions that will utilize the pushing/popping helpers but at the same time providing something easier to understand for people new to the EnvBuilder.

Reviewed By: mvitousek

Differential Revision: D28550336

fbshipit-source-id: 83d8717099932b74501572b8fa46368d95b527bd