fix locationless this type crash on method unbinding
Summary:
The error added to warn against implicit method unbinding via function subtyping in

https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbcode/flow/src/typing/subtyping_kit.ml?commit=cdde4cb0ce6c1b179e9d0b7eb0c1f6a3d6986a45&lines=1225-1229

references the `reason` of the lower bound function's `this` param. Because this location is included in the output of a `EMethodUnbinding` error, this means that all `this` params must have locations. Previously, when creating dummy `this` types, we would use locationless reasons under the assumption that the dummy types would never make it into an error message. This was the case up to the change adding this new error, so now we must include a location for all `this` parameters. Here we add these locations.

The test added is a repro of this bug that creates a locationless `this` param by importing from another file.

Reviewed By: samwgoldman

Differential Revision: D29174341

fbshipit-source-id: e4a41f084061b88c5d67826c511596bacd0b93ab