Represent shmem addrs as byte offsets into a char bigarray
Summary:
In the shared heap, addresses are stored as byte offsets from the heap base
pointer. Before this diff, when representing an address in OCaml, we would
convert this byte offset into a word offset. This diff removes the conversion,
so we also represent addresses as byte offsets in OCaml.

To make this work, I changed the word-based Bigarray to be char-based instead.
The parameterization of the Bigarray affects the get/set operations as well as
the size of the data returned, similar to changing a C array from int* to char*.

Because we still want to read/write words, just at char offsets, I changed from
using the get/set API to primitive functions, which do exactly what we want.

Note that we still represent sizes in word units, so we still need to take some
care when computing a new address from an existing address (byte offset) and a
size (in words).

The `(char, int8_unsigned_elt, c_layout) Array1.t` type is a more generally
useful type as well. It is sometimes called a "bigstring" and there are
libraries that provide useful functionality for it, notably Bin_prot.

Differential Revision: D28401827

fbshipit-source-id: f2eaa9b19682beb205fd02fe9a0dca8cdc566335