[lsp][extract] Find a list of locally undefined variables of extracted functions in target scope
Summary:
This diff continues the work from D29140311, and use the `(def * scope) list` generated from the last diff to decide whether a variable used inside the extracted statement will be undefined after extraction.

The implementation is mostly straightforward using the idea I mentioned in the last diff:

> For variable `v` unused inside the statements, we know its def `d` and scope `s`. If the def happens completely inside the range of selected statements, then it is defined. Else, we see the scope `s'` where we place the extracted function. The the variable `v` is undefined iff `s' > s`.

To implement the check `s' > s`, I added a helper function to the `Scope_api` to compute whether two scopes have the `>` relation. There are also some complications due to the fact that there might be two scopes associated to the same loc, which I explained my strategy in the inline code comment.

Reviewed By: gkz

Differential Revision: D29113791

fbshipit-source-id: a31dfb47d54418063dc562aaeb5e24fb4f312b34