[lsp][extract] Stop exposing and testing `create_extracted_function`
Summary:
I originally envisioned this function to contain the implementation of adding undefined variables as parameters and returns. However, at that time I did not realize that the list of undefined variables are actually context sensitive, and cannot be derived from the statements alone. Therefore, the abstraction of this function is not very useful, and could just exist as an internal helper function alone.

For testing, the implementation of this function will be covered by other integration tests anyways, so it's also not very useful to test this alone.

For the above reasons, I removed this function from interface and tests.

Reviewed By: gkz

Differential Revision: D29144993

fbshipit-source-id: 622de610796afdf54d6ea2360a58f4bff5aa0f1b