[lsp][extract] Find a list of variables definitions with their scopes that are used inside extracted statements
Summary:
## Background

This diff starts the first step of computing the "boundary" of the extracted function. That is: finding a list of variables that are undefined within this extracted function, and the new variables that are defined inside the function that needs to be returned to the outside world.

Sitting at the core of the task is a problem of finding an info list of variables that are used within the extracted statements. More importantly, we need to know the place of variables' definition and the scope of definition. i.e. we need a `(def * scope) list`.

To see how that information is needed, we see how we can solve the following two problems with this list:

1. Finding variables that are undefined.
For variable `v` unused inside the statements, we know its def `d` and scope `s`. If the def happens completely inside the range of selected statements, then it is defined. Else, we see the scope `s'` where we place the extracted function. The the variable v is undefined iff `s' > s`.
2. Finding variable defined inside the statements.
We look at every def in the list, and then only keep the def where the loc of the def is within the range of the extracted statements.

## Implementation

I first build the scope information using `Scope_builder` that is already used to power find-def and autocomplete LSP services.
However, the scope information returned from the builder is not directly and efficiently usable for the function extraction use case, so I have to make two passes over the information to build the list.
The first pass looks at the `use_def_map` in the `locals` fields of every scope. We collect the defs into a set that are used within the range of extracted statements.
The second pass then looks like the def set in every scope, and associate every def found in the first pass with a scope object.

Reviewed By: gkz

Differential Revision: D29140311

fbshipit-source-id: f5270d8347deb7eb93c69489e3065e8a90552b7f