Make handling of missing aloc_tables more precise
Summary:
When abstract locations was originally designed, we only stored aloc tables when
the abstract_locations configuration was enabled. Under types-first 2.0, we
always store the table in the heap, and store signatures with offsets into the
table. The configuration instead controls whether we create types using concrete
or keyed alocs.

Because of this, calls to get_aloc_table_unsafe should always succeed, even if
we don't use the result. The one exception is for check-contents, which still
needs to handle the case of a missing heap entry.

This diff moves the exception handler to a more specific code path and adds a
detailed comment explaining why the handler is necessary, and what future work
is required to remove it.

Reviewed By: panagosg7

Differential Revision: D29233008

fbshipit-source-id: 4729124be7e53c93bae9867eb80c64e8ef3c8f9f