[new-env] Don't reorder to happen after annotated providers
Summary: The way that we break dependency cycles is to add annotations, but annotations didn't really matter previously. In order to support breaking cycles, this diff adds information to the provider analysis about whether a variable is annotated or not (for functions, this is determined by whether their return types are annotated). When building an ordering, we exclude locations that are depended on if they correspond to annotated variables--we don't need to check such a variable's declaration before we check something that uses it, because we can have preloaded the type during statement_decl.

Reviewed By: jbrown215

Differential Revision: D28145236

fbshipit-source-id: 04dc1152e5a602d06e8c9c996d98684435883610