[lsp][extract] Simplify arguments passed into `new_function_call_replacer`
Summary:
It turns out that we do not have to pass in 3 parameters into `new_function_call_replacer` to correctly find where to place the `newFunction()` call and where to delete.

Passing the union is already enough. We can first use the union of all locations to figure out what are the statements we need to touch (outer if-else), and then use the check of `Loc.start statement_loc == Loc.start extracted_statements_loc` to see where is the first statement (inner if-else).

Reviewed By: gkz

Differential Revision: D29151128

fbshipit-source-id: c81eebb17f5c544ed96302b24d2838ed946ab09c