[lsp][extract] Merge `new_function_call_replacer` and `insert_new_function_in_body_mapper` into `refactor_mapper`
Summary:
We want the code that replaces the original statements with `newFunction()` call to contain a correct list of variables as arguments. This requires us to pass `undefined_variables` to the replacer.

Since the undefined variables list is already passed to the code that appends `newFunction` declaration, we might just as well pass the undefined variables list to a single function, and make the function does the job of appending declaration and replacing the original statement list with `newFunction()` call in a single pass. This improves performance, and also makes the flow of data clearer in this file.

Reviewed By: gkz

Differential Revision: D29151140

fbshipit-source-id: 79b3f577ccb8e21e6c3151faeb23cc8ddcb06fc3