[lsp][extract] Unify ways to insert `newFunction` declaration
Summary:
Right now we are inserting the `newFunction` declaration in two different ways: a special case for toplevel insert and a ast-mapper-based approach for inner function insert.
This kind of structure is blocking some refactors I am planning (passing in undefined variables to `newFunction()` call, see diffs later in stack), so I refactored them to use the same ast-mapper-based approach in this diff.

Reviewed By: gkz

Differential Revision: D29151116

fbshipit-source-id: 328664982fe63079e2623d2450371c1567814f56