[indexed-access] When index is string literal, use PropertyType destructor rather than ElementType
Summary:
Indexed Access currently always uses `ElementType` under the hood. However, there is a difference between `$ElementType<T, 'k'>` and `$PropertyType<T, 'k'>` when `T` is an instance. This means we cannot codemod `$PropertyType<T, 'k'>` when `T` is an instance to the indexed access type syntax.

`GetElemT` does not behave equivalently with instances and objects:
https://flow.org/try/#0CYUwxgNghgTiAEkoGdnwMLwN4Ch7wCNYAueAOwFcBbAkGAbhwF8cAKARlIBIBRCEKiDIAXACoBPAA4gAPOgA08AOREYSgHwBKevBw5hUhAHl4AXmx5CJctVoNmbTvF79BIidJlHFK2Bu16oEhw8ABusIik6IxB0CHhMPAA9qRGjDhgANq+agC6jEnZqkr5QA

Here, I change indexed access so that it uses `$PropertyType` when the index is a string literal. This produces the behaviour we want.

Temporarily, the type normalizer will print `$PropertyType<T, 'k'>` for `T['k']`. This is fixed in a diff up the stack.

In the future, I will also update optional indexed access with the same fix (it's a bit more complicated).

Reviewed By: panagosg7

Differential Revision: D29221381

fbshipit-source-id: 85461645788e6f2f92d8ea0adccdce5f02c08960