[indexed-access] Fix printed type for indexed access when PropertyType destructor is used
Summary:
Temporarily add a flag to the `PropertyType` destructor so when its origin is the indexed access type syntax, indexed access is printed by the type normalizer. This is similar to the existing flag in the `ElementType` destructor.

In the future, we will eliminate the `$PropertyType` and `$ElementType` type utilities, and only have the indexed access syntax. When this happens, these flags will be removed.

Reviewed By: panagosg7

Differential Revision: D29257501

fbshipit-source-id: 8031178e718f1a35a69010f4cea138c4f111016d