Parse internal comments in enums
Summary:
This diff fixes a bug where comments after the last enum member, called internal comments in the parser, are not in the parser or printer. I copied most of the implementation here from internal comments in objects. I also fixed a bug where the unknown members ellipses was ignored in defaulted string enums.

A couple caveats as it was tricky to print the internal comments along with the unknown members ellipsis in the same place as the original code:
- The printer will always put internal comments before the ellipsis
- The printer adds an extra newline after comments. This is due to the `Indent` needing a newline to force unindent but the comment containing its own newline.

Reviewed By: gkz

Differential Revision: D29167570

fbshipit-source-id: 6dcab3790e726730285d5f3a6a10ca9535d27eb6