Add support for enums to the AST differ
Summary: This diff adds support for enums to the AST differ. Without enums in the differ, codemods that run over files with enums will replace the entire enum even though nothing has changed. This is usually okay, but small changes with the pretty printer cause changes in the output. By avoiding replacement if nothing is changed, we minimize the chance of unrelated changes when making codemods.

Reviewed By: gkz

Differential Revision: D29285028

fbshipit-source-id: 2de05323a41ccc43cb6816171d6ef19a8b3a3760