[lsp][extract] Collect all escaping definitions inside extracted statements
Summary:
The final step to produce well-typed code for function extraction code action is to add returns for variables defined inside the extracted statements.

We could iterating through all the defs in the top-level statements within the selection range, but that would over-approximate the number of returns needed. For example, we can consider this program:

```
const a = 1; // selected
const b = a + 1; // selected
console.log(b);
```

If we collect all the defs, then `a` and `b` both have to be returned, but the rest of the program does not use `a` at all.

A better and much simpler approach is to only find defs that are "escaping" from the selection range. The criterion for escaping is also very simple:

> A def is escaping if it's defined inside extracted statements but have at least one use outside of extracted statements.

To collect such defs, we only need to make a one-pass visit over all the use-def map from scope information to collect all of them into a set.

Reviewed By: gkz

Differential Revision: D29171316

fbshipit-source-id: ee67da00ea425f3fbbda4db0d163bb2704c99ac6