[ast-differ] Record parent of expression diffs
Summary:
Currently, we always print parentheses around expressions except for a few special cases. The fundamental issue is that we don't have context about the parent of the expression being diffed, so we don't know whether parenthesis is necessary.

This diff is the first step to implement better pretty printing for expression diffs. I make the expression diff record not only the node itself but also its parent. This should gives the diff printer enough information to decide whether to print parenthesis. It also gives the diff printer more information of the context to make better heuristics for printing lists.

Reviewed By: gkz, vrama628

Differential Revision: D29237814

fbshipit-source-id: 435a7b6f2452cd4b90de791a867733365f7b1537