[ast-differ] Avoid wrapping parentheses around expression if the parent doesn't require parentheses
Summary:
Based on the parent information collected in the previous diff, we can now avoid printing a lot of unnecessary parenthesis now.

To keep things simple in this diff, I only updated the case where it's parentheses are never required. The case when the parent of expression is expression or a spread, where precedence needs to be carefully considered, is not implemented yet.

Reviewed By: gkz

Differential Revision: D29242043

fbshipit-source-id: 2ac298c527294835f37a3856ce1b28d4167e0856