Simplify aloc table representation
Summary:
Before this diff, the ALoc module dealt with the construction of aloc tables as
well as their use. Since we shipped new signatures, aloc tables are now
constructed by the type sig parser.

This diff is a step towards simplifying the APIs here. We now use a sorted
array of concrete locations as the aloc table. Because the array is sorted, we
no longer need a reverse table, since we can do binary search instead.

I will follow up with more cleanup here, since the ALoc module's API is still
pretty messy.

Reviewed By: panagosg7

Differential Revision: D29119262

fbshipit-source-id: cf882f3d6e77a51ddf10ea91ee4046b98f9f1eda