Store docblocks in the new heap
Summary:
Before this change, each file's docblock was stored in shared memory as a blob
and registered with the hash table. This diff still stores the data as a blob,
bu instead of using the hash table, adds a pointer to the blob from the "checked
file" heap object.

The docblock is kept alive by the file to which it belongs, and can be found by
file name exactly as before: by finding the checked file object, then
indirecting to the docblock.

I am planning to submit many similar diffs, affecting other file-related data
which is stored under multiple keys currently. This will help our shared hash
table continue to scale as the size of the repo grows.

This also simplifies the transactional behavior of incremental checks. Currently
each hash table entry is "oldified" and potentially "revived" at the end of a
recheck. As we move more of these hash table entries to be indirected from the
file heap object, we can oldify/revive only one key per file.

Furthermore, the mutator/reader APIs can be simplified once this migration is
complete. These APIs are necessary to ensure that code is consistently reading
from the "old" or "new" hash table keys while a recheck is ongoing. Instead, we
should be able to pass a file address, where everything reachable from that
address is consistent.

Reviewed By: panagosg7

Differential Revision: D29188242

fbshipit-source-id: 0af8a9cc856d6e7939fc8862fb3beffc5ca9b475