[lsp][extract] Provide no extraction refactors when user selection includes control flow statements
Summary:
When we encounter `return`, `break`, `continue`, `yield`, `label`, rewrites are not straightforward in general, since they contain non-local behaviors. Therefore, it's better to provide no refactors at all in these cases, instead of providing a wrong/confusing one.

This diff simply detects these statements and gives up when we see any single one of them. In the future, we could relax the handling for break and continue a little bit, if they are inside a loop, and they are not break/continue with labels.

Reviewed By: mvitousek

Differential Revision: D29306797

fbshipit-source-id: 8ddfd65d0f757fed1822eae556c23307d2474f3d