[lsp][extract] Locate the single scope to put `newMethod` definition
Summary:
The first step of support extract to method feature is to figure out where to insert the new extracted method. Unlike extracting to a normal function, extracting to method brings more constraints:

- Methods must live inside a class body, instead of any functions or toplevel scope.
- Methods must be called by `this.newMethod()`.

Since our goal is to ensure that the resulting code does not result in type error, we must make sure that `this.newMethod()` is a valid call. According to JS behavior, `this.newMethod()` will call the method within the same inner-most class that contains the `this.newMethod()` call. Therefore, there is only one valid class scope to put the `newMethod` declaration: the smallest containing class of the extracted statements.

This diff creates another visitor to compute such scope. Later diff will use this scope to insert methods.

Reviewed By: mvitousek

Differential Revision: D29343968

fbshipit-source-id: bd537cac9c386762bf2ffc016543d44c13de52c3