preserve interface-ness through read-only utility
Summary:
type utilities that make use of `ObjKit` did not preserve the "interface-ness" of their input, as the object kit types would construct an object for the output every time.

This threads the "interface-ness" of the utility input through the object kit operations so that we can reconstruct the output properly.

Reviewed By: jbrown215

Differential Revision: D29202142

fbshipit-source-id: 54a23ecedba055835f24c864d6e39d615311a2b9