[lsp][extract] Detect existence of `this` and `super` within extracted statements
Summary:
When we have `this`/`super` inside the extracted statements, we can no longer extract the statements into a normal function. Instead, our only option is to extract them into a method and then call `this.newMethod()`.

This diff detects the existence of `this`/`super`, and then temporarily provide an empty list of extract refactors. In later diff when extract to methods have been implemented, it will be replaced by extract to methods refactors.

Reviewed By: mvitousek

Differential Revision: D29344487

fbshipit-source-id: 9b43b396df11d8ad40bcc51c612daf6893c2ae2f