[glean] Move glean indexer SourceOfExport fact generation off of file_sig with_exports_info
Summary: We have a better-engineering goal of moving off of the `file_sig_with_exports_info` data structure. The code that generates `SourceOfExport` facts in the Flow Glean indexer was using that data structure. This diff rewrites that code to use the `Type_sig` data structure instead. After this diff, the only part of the Glean indexer that still uses `file_sig_with_exports_info` is the part that generates `SourceOfTypeExport` facts, which I'll rewrite in a similar way in a future diff.

Reviewed By: samwgoldman

Differential Revision: D29310638

fbshipit-source-id: 3df545a4a989aac8df5db85eac3c320c71387202