Move SourceOfTypeExport fact generation off of file_sig_with_module_info
Summary: completes the moving of the glean indexer off of `file_sig_with_module_info`. Also removes the necessity of the `type_declaration_map`s that were being passed around in the indexer, which is great because those were ugly. The next diff will remove the `type_declaration_map` stuff from the code.

Reviewed By: samwgoldman

Differential Revision: D29397235

fbshipit-source-id: e490a6a346191dfe7b2bd6c2f8a8d766694c48a0