Drop exports info from file sigs
Summary:
The logic to extract "exports info" as part of the file sig analysis was
initially added for types-first, which was recently obsoleted by new signatures
based on type sigs.

In the intervening time, some other code started depending on this information,
but we've now updated all those dependents to no longer require the exports
info, so this can finally be removed.

Reviewed By: vrama628

Differential Revision: D29399343

fbshipit-source-id: 8b5fc8e7afbe15a3fbcce37099289834fcb53c07