abstract union_find.ml implementation
Summary:
I'm planning to reuse the union-find implementation for the "annotation" graph.
This diff moves the relevant functions from context.ml to union_find.ml.

Functions like `find_root` now also return the updated graph, the storing of which
is now the client's responsibility (in this case context.ml).

In type.ml, the `Constraint` does not need to be part of the recursive module
definition that includes `TypeTerm` and so on, since the latter do not depend on
the former. This helps remove some of the redundancy in type alias definitions
that we earlier had.

Reviewed By: samwgoldman

Differential Revision: D28493428

fbshipit-source-id: aa901d5d4450428bd09a6cbca8070a75d717128a