while loops
Summary:
This handles refinements introduced by while loops. Specifically, inside the loop we keep the refinement introduced by the guard in scope. After the while loop we negate the guard's refinement and apply that to the _name_ of the variable instead of the ssa_id. No matter what writes happen inside the loop (other than a break), the only way to exit to the next line would be to negate the guard.

We do not follow the same Unresolved/assert_ssa_env strategy that the ssa_builder uses. This approach, while more complete, can extremely easily introduce cycles in the dependency graph for values, which the future ordering analysis will not be able to handle nicely. In order to minimize those cycles we instead:
1. Determine which variables were _written to_ inside the loop. We do this by grabbing the WriteSets of each value, looking THROUGH refinements. If the write set of a variable _after_ the loop is different than before, we can conclude that the variable was written to inside the loop. Furthermore, even though we model refinements as writes, since we look through refinements to find these write sets we won't count refined values as values that are written to.
2. Reset our state back to what it was before reading the loop guard and start again. Any variables that are written to inside the loop are havoced to break the potential cycles those writes would create.
3. At the end of the loop, we deeply remove refinements associated with the loop guard. These refinements can end up in PHI nodes that are created due to abnormal control flow (see the added test case mentioning control-flow writes), and we want to make sure they are not accessible at the end of the loop. After removing those refinements, we merge the post-loop state with the pre-loop state to represent either entering the loop or bypassing it completely. If there was no break inside the loop then we negate the loop guard's refinement and apply it to the name of the variable that was refined.

I broke out as much as I could into helper functions that we can re-use across all the loop implementations, which should drastically simplify the "client code" in the loop implementations themselves.

Reviewed By: mvitousek

Differential Revision: D28265406

fbshipit-source-id: d8548ab7db0a3386cb1c4ca7e1ef5afc96bb73da