fix issue where references to this inside of nested objects were accidentally permitted
Summary:
In an attempt to be more efficient when checking for `this` inside of object methods, Flow was terminating its search at every point where it encountered an object literal, since the idea was that every object literal was checked anyways, so looking inside of nested ones would be redundant.

In practice, this meant that place where `this` was used in object properties that were not methods were ignored, and since `this` refers to the enclosing function, if `this` was used in an object property inside of a method, we would allow an unsafe usage (see added test for an example of this).

This removes the short-circuit case from the `this`-finder, meaning we now correctly find these usages.

Reviewed By: jbrown215

Differential Revision: D29491330

fbshipit-source-id: d07a8b03ed87851f9d6694785249fa34503fd7da