Fix issue with name collisions in declared modules
Summary:
This diff fixes an issue where declarations in a `declare module` would be lost if the declaration's name is the same as a declaration that was visited previously in a different libdef. This is intentional when both declarations are at the top level of a libdef**, but it shouldn't apply to declarations enclosed within a `declare module`, which should be distinct from other declarations within libdefs.

This was modeled by the `exclude_symbols` code in the environment, and this diff fixes it by simply resetting the `exclude_symbols` set when we enter a declared module and restoring it when we pop out of it.

Fixes https://github.com/facebook/flow/issues/8604

Reviewed By: mroch

Differential Revision: D29498984

fbshipit-source-id: feb7fcde5b861f1148a5ad13ab5eae276af794c2