[ast-differ] Track the new parent of the changed expression node
Summary:
The goal of keeping track of the parent of the expression is to make the pretty-print of diff aware of the context. These context can inform the pretty printer about whether to add parenthesis or whether to coalesce several different edits into one.

Right now we are keeping track of the parent from the old AST. Although in principle you can reconstruct the parent of the new AST using the expression node information, it is not very convenient. I assert that in general parent from the new AST node will be more useful for the following two reasons:

- The diffing process is used to inform us what should be done to get a new AST. The old AST is already irrelevant as this point, since all we care about is the location of the change.
- The pretty printer, depending on the context, might want to print a larger diff to keep the printed diff clean. In this case, the new parent node would come in handy since it can be directly used for printing.

Therefore, I decided to switch to track the new parent node in this diff. In the next diff, I will the first use of it to improve diff pretty-printing.

Reviewed By: gkz

Differential Revision: D29373156

fbshipit-source-id: e8e5c5c6ddc5b4d01a2f76d6f43bcd815933f3a4