[lsp][extract] Implement method extraction
Summary:
This diff connects the work of the previous few diffs to implement extraction for methods. At the end, we will be able to select statements containing this inside classes and refactor them from

```
class A {
  test1() {
    this.test1(); // selected
  }
}
```

to

```
class A {
  test1() {
    this.newMethod();
  }
  newMethod() {
    this.test1();
  }
}
```

It uses the `find_closest_enclosing_class_scope` function introduced in D29343968 (https://github.com/facebook/flow/commit/1250a80a857aa705656064878b28ced946d052f2) to compute the scope to put in the method definition, and uses the existing function to generate a method call with all the parameters and return values correctly passed along. The AST builder functions `create_extracted_function` and `create_extracted_function_call` are slightly tweaked to support methods.

It turns out that this code can be easily extended to support statements that do not contain `this` as well, so I also added `extract_to_method_refactors` to the not in-class case.

Reviewed By: vrama628

Differential Revision: D29366101

fbshipit-source-id: f3444d6f8493f92aec73aeeed2cd4da2d1d2805f