[lsp][extract] Make `create_extracted_function_call` return a list of statements
Summary:
This diff continues the work to correctly refactor statements that contain assignments. For example,

```
function test() {
  let fooo = 3;
  const a = 3; // selected
  fooo = a + 2; // selected
  console.log(a + fooo)
}
```

should be refactored into:

```
function test() {
  let fooo = 3;
  let a;
  {a, foo} = newFunction(fooo);
  console.log(a + fooo);
}
function newFunction(fooo) {
  const a = 3; // selected
  fooo = a + 2; // selected
  return {a, foo};
}
```

Note that this requires us to replace the selected statements with zero or more `let` definitions before the `newFunction()` call. Therefore, this diff does the preparation work to make the replacement works with a list of statements.

Reviewed By: vrama628

Differential Revision: D29440591

fbshipit-source-id: 1b935a8f63a370d8315f1c35ad583fb52fc21705