[lsp][extract] Collect a list of assignments of definitions that would be shadowed after refactor
Summary:
This diff starts to tackle the problem of incorrectly shadowed assignments after refactor. To understand what the problem is, we can look at the following example:

```
function test() {
  let fooo = 3;
  const a = 3; // selected
  fooo = a + 2; // selected
  console.log(a + fooo)
}
```

is currently refactored into

```
function test() {
  let fooo = 3;
  const a = newFunction(fooo);
  console.log(a + fooo);
}
function newFunction(fooo) {
  const a = 3; // selected
  fooo = a + 2; // selected
  return a;
}
```

This behavior is wrong since `fooo` is shadowed inside `newFunction`, so the reassignment to `fooo` does nothing.

In general, the problem is caused by variables `v` such that all of the following is true:

- `v` is used outside of extracted statements
- `v` is initially defined outside of extracted statements
- `v` has a write inside extracted statements

This diff finds all such variables. In later diffs we will use these variables to perform correct refactors.

Reviewed By: vrama628

Differential Revision: D29436699

fbshipit-source-id: e9676caab05a9358308dff1fe0b8cb908975155e