[lsp][extract] Use `let` declaration if one of the escaping variables are mutated outside of extracted statements
Summary:
Right now, we are always using `const` declaration to collect the escaping definitions inside extracted statements. However, this does not always type check. Consider the following example:

```
let a = 3; // selected
a = 2;

// incorrect refactor
const a = newFunction();
a = 2; // oops, assign to const
function newFunction() {
  let a = 3;
  return a;
}

// correct refactor
let a = newFunction();
a = 2;
function newFunction() {
  let a = 3;
  return a;
}
```

This diff combines the scope information and SSA information to decide whether we need to use `let` declaration. More specifically, if a locally defined escaping variable has a write outside of extracted statements, then we must use `let` so that it can be correctly reassigned.

Reviewed By: jbrown215

Differential Revision: D29474557

fbshipit-source-id: d97dc1078621eddd38adabe920750b001552adbe