move babel_loose_array_spread default out of flowConfig
Summary:
when we store defaults in `flowConfig.ml`, we can't tell whether the user explicitly set the value to the default, or if it was unset. This makes it hard to reformat the flowconfig.

Instead of storing `bool`, we should store `bool option` and move the default to `commandUtils.ml` where we construct `Options` by combining the CLI flags, flowconfig, and defaults

Reviewed By: mroch

Differential Revision: D29614581

fbshipit-source-id: cf33a3c968141d6640cd8bb8f59c1a6c8f9928b3