[lsp] Make `Autofix_imports` generate import changes directly from added import statements
Summary: This diff removes the last blocker to use the `Autofix_imports` API to generate diff to add imports from `remote_converter`. Currently, the remote converter API returns a list of import statements inside of raw information. This diff adds another method to output the imports in the format of `Autofix_imports` and adds a new function in `Autofix_imports` to batch add imports.

Reviewed By: vrama628

Differential Revision: D29665938

fbshipit-source-id: 8b9af7a3243a80f8deec95f5b7e71b598768d475