[lsp][extract] Add type annotations to `newFunction` parameters
Summary:
This diff bases on previous diffs' work to insert type annotations for function parameters. It uses the `TypeSynthesizer` API in the last diff to synthesize type annotations based on location and add imports automatically.

This diff has not addressed the scoping issues of generic type parameters of inserted types yet. It will be addressed in the next few diffs.

Reviewed By: vrama628

Differential Revision: D29640143

fbshipit-source-id: a2799d06cab20bc45bcc2ab57e4e3d4fc41a5201