[lsp][extract] Fix incorrect undefined variable detection when toplevel only contains one function
Summary:
This diff fixes a very subtle bug that results in fewer undefined variables being reported for extract to function, but it's fundamentally caused by incorrect assumption made when calling `Scope_api.scope_of_loc`.

## Problem

`Scope_api.scope_of_loc info loc` returns all the scopes that have the specified `loc`. This works well in almost all cases, except in one case when we have a program that only contains 1 toplevel function. For example:

```
function test(a: number) { console.log(a) }
```

In this case, 3 scopes will have the same loc that contains the entire AST:

1. Hoisted scope of toplevel that contains def of `test`. (Always have `scope_id = 0`)
2. Lexical toplevel scope that contains nothing.
3. The function parameter scope that contains def of `a`.

The intention is to only collect the first two scope, but the third scope sneaks in, which makes the program incorrectly believe that `test` and `a` are both defined in toplevel scope. Therefore, when we are trying to extract `console.log(a)` into the toplevel scope, we incorrectly think that `a` will not become undefined.

Note that extract to inner function does not suffer from the same problem, since their scope location is defined by the range between braces. Even the inner function scope only have one function inside, the locs will still be different due to additional `{ ... }` of the outer scope.

## Solution

The above analysis shows that the problem only exists for toplevel scope, so we can continue using `Scope_api.scope_of_loc` for extract to inner functions. However, for toplevel scope, we need a more specialized function that does not suffer from the same problem. I created `Scope_api.toplevel_scopes` function that only finds the first two scopes in the above section and nothing else.

Reviewed By: vrama628

Differential Revision: D29746012

fbshipit-source-id: c81305e06689c5d358a1d4a56b5a7dc9b5f39176